change parameter data type
arduino.py will now send signed integers to arduino when sending parameters as is consistent with the Python datatype. Previously, the python int32 was being packed as uint32, achieving confusion and limiting the ability to send negative values as parameters. 

If uint32 capacity (>2e9) is required, we should add a feature where python is using tables.UInt32Col datatype and then is sending the same to arduino.